Add preparation stage to environment provider.
If a proper preparation step is provided, the environment provider
will attempt to prepare the IUT according to the JSONTas instruction.
More information will be added to documentation.

Also added the artifact events, activity event and tercc event
to dataset for JSONTas providers to use.
Added a check and get for an artifactPublished from GraphQL API.
ArtifactPublished can be None.

Updated to latest JSONTas which added features required for the
preparation to work.