Set a specific python version and update so that pylint is happy
Pylint complained about not using f-strings.
In order to add f-strings we would have to stop supporting py2
so I set a fixed python version to the one used in the Dockerfile
and reformatted all files.
Note that the logging module calls are not updated because of
how the logging module wants to handle it its own way.