Raise exceptions from the environment provider
When running v0 ETOS the environment provider would not raise
exceptions back to the request_environment method in the ESR.
This caused the ESR to raise a generic Exception with the error
text instead.
The problem here is that the *CheckoutFailed exceptions can
sometimes contain an empty string and the ESR would raise
Exception('') which tells us nothing about what happened.
This problem also exists for v1 ETOS where we attempt to write
the error to termination-log, but we will write an almost
completely empty traceback that says nothing.

Created a new method that the ESR can call to have exceptions
thrown up the stack.