Get traceparent from environment when running in controller env
When we are running in the controller environment we don't use the
environment provider as a library and won't automatically get the trace
parent from the suite runner.
We now assume that the ETOS controller will pass the traceparent to the
environment provider job as an environment variable and load it.

This will mean that there's a slight difference in the trace produced
from the controller environment vs the v0 environment, but I don't see a
big problem with that since the controller environment does do the
environment checkout a bit differently and that will be reflected in the
trace.